requirements.txt
streamlit 
snowflake-connector-python 
pandas 
trulens-eval 
openai